refactor: useReciever, append,remove 활용
