+++ src/pages/Order/Reciever/RecieverModal.tsx
@@ -1,6 +1,7 @@
-import { useState } from 'react';
 import { ErrorContainer } from '@/styles/Login.styles';
 import type { RecieverType } from '@/pages/Order/Order';
+import type { UseFieldArrayRemove, UseFieldArrayAppend } from 'react-hook-form';
+import type { FormValues } from '@/pages/Order/Order';
 import {
   ModalBackdrop,
   ModalBox,
@@ -20,89 +21,33 @@ import {
   ModalListScroll,
   ModalBtnRow,
 } from '@/styles/Order/OrderModal.styles';
+import useReciever from '@/hooks/useReciever';
 
-interface RecieverModalProps {
+type RecieverModalProps = {
   open: boolean;
   onClose: () => void;
   onComplete: (list: RecieverType[]) => void;
   initialList: RecieverType[];
-}
-
-type FieldError = {
-  name?: string;
-  phone?: string;
-  count?: string;
+  append: UseFieldArrayAppend<FormValues, 'reciever'>;
+  remove: UseFieldArrayRemove;
 };
 
-function RecieverModal({ open, onClose, onComplete, initialList }: RecieverModalProps) {
-  const [list, setList] = useState<RecieverType[]>(initialList);
-  const [fieldErrors, setFieldErrors] = useState<FieldError[]>([]);
-
-  const handleChange = (idx: number, key: keyof RecieverType, value: string | number) => {
-    setList((prev) => prev.map((item, i) => (i === idx ? { ...item, [key]: value } : item)));
-    setFieldErrors((prev) => {
-      const next = [...prev];
-      if (!next[idx]) next[idx] = {};
-      let err = '';
-      if (key === 'name') {
-        if (!String(value).trim()) err = '이름을 입력해주세요.';
-      } else if (key === 'phone') {
-        if (!String(value).trim()) err = '전화번호를 입력해주세요.';
-        else if (!/^01[016789][0-9]{3,4}[0-9]{4}$/.test(String(value).replace(/-/g, '')))
-          err = '올바른 전화번호 형식이 아닙니다.';
-        else {
-          const phones = list.map((r, i) => (i === idx ? value : r.phone));
-          if (phones.filter((p) => p === value).length > 1)
-            err = '동일한 전화번호는 입력할 수 없습니다.';
-        }
-      } else if (key === 'count') {
-        if (!value || Number(value) < 1) err = '수량은 1개 이상이어야 합니다.';
-      }
-      next[idx] = { ...next[idx], [key]: err };
-      return next;
+function RecieverModal({
+  open,
+  onClose,
+  onComplete,
+  initialList,
+  append,
+  remove,
+}: RecieverModalProps) {
+  const { newList, fieldErrors, handleChange, handleAdd, handleRemove, handleComplete } =
+    useReciever({
+      open,
+      onComplete,
+      initialList,
+      append,
+      remove,
     });
-  };
-
-  const handleAdd = () => {
-    setList([...list, { name: '', phone: '', count: 1 }]);
-    setFieldErrors([...fieldErrors, {}]);
-  };
-
-  const handleRemove = (idx: number) => {
-    setList((prev) => prev.filter((_, i) => i !== idx));
-    setFieldErrors((prev) => prev.filter((_, i) => i !== idx));
-  };
-
-  const handleComplete = () => {
-    const errors: FieldError[] = [];
-    let hasError = false;
-    const phones = list.map((r) => r.phone);
-    list.forEach((item, idx) => {
-      const err: FieldError = {};
-      if (!item.name.trim()) {
-        err.name = '이름을 입력해주세요.';
-        hasError = true;
-      }
-      if (!item.phone.trim()) {
-        err.phone = '전화번호를 입력해주세요.';
-        hasError = true;
-      } else if (!/^01[016789][0-9]{3,4}[0-9]{4}$/.test(item.phone.replace(/-/g, ''))) {
-        err.phone = '올바른 전화번호 형식이 아닙니다.';
-        hasError = true;
-      } else if (phones.filter((p) => p === item.phone).length > 1) {
-        err.phone = '동일한 전화번호는 입력할 수 없습니다.';
-        hasError = true;
-      }
-      if (!item.count || item.count < 1) {
-        err.count = '수량은 1개 이상이어야 합니다.';
-        hasError = true;
-      }
-      errors[idx] = err;
-    });
-    setFieldErrors(errors);
-    if (hasError) return;
-    onComplete(list);
-  };
 
   if (!open) return null;
 
@@ -120,14 +65,14 @@ function RecieverModal({ open, onClose, onComplete, initialList }: RecieverModal
           <ModalAddBtn
             type="button"
             onClick={handleAdd}
-            disabled={list.length >= 10}
-            style={{ opacity: list.length >= 10 ? 0.5 : 1 }}
+            disabled={newList.length >= 10}
+            style={{ opacity: newList.length >= 10 ? 0.5 : 1 }}
           >
             추가하기
           </ModalAddBtn>
         </ModalAddBtnRow>
         <ModalListScroll>
-          {list.map((field, idx) => (
+          {newList.map((field, idx) => (
             <ModalInputRow key={idx}>
               <ModalInputTitle>
                 <div>받는 사람 {idx + 1}</div>
@@ -174,7 +119,7 @@ function RecieverModal({ open, onClose, onComplete, initialList }: RecieverModal
             취소
           </ModalActionBtn>
           <ModalCompleteBtn type="button" onClick={handleComplete}>
-            {list.length}명 완료
+            {newList.length}명 완료
           </ModalCompleteBtn>
         </ModalBtnRow>
       </ModalBox>
